# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Airdrop Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /airdrops/{id}/tasks:
    get:
      operationId: listAirdropTasks
      tags: [airdrop-tasks]
      summary: List tasks for an airdrop
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AirdropTask"
    post:
      operationId: createAirdropTask
      tags: [airdrop-tasks]
      summary: Create a task for an airdrop
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AirdropTaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AirdropTask"
  /airdrop-tasks/{id}:
    patch:
      operationId: updateAirdropTask
      tags: [airdrop-tasks]
      summary: Update an airdrop task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AirdropTaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AirdropTask"
    delete:
      operationId: deleteAirdropTask
      tags: [airdrop-tasks]
      summary: Delete an airdrop task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    AirdropTask:
      type: object
      required: [id, airdropId, title, taskType, isRequired, sortOrder, createdAt]
      properties:
        id:
          type: integer
        airdropId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        link:
          type: ["string", "null"]
        taskType:
          type: string
          enum: [follow_twitter, join_telegram, join_discord, complete_quiz, stake, bridge, swap, provide_liquidity, register, other]
        isRequired:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
    AirdropTaskInput:
      type: object
      required: [title, taskType]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        link:
          type: string
        taskType:
          type: string
          enum: [follow_twitter, join_telegram, join_discord, complete_quiz, stake, bridge, swap, provide_liquidity, register, other]
        isRequired:
          type: boolean
        sortOrder:
          type: integer
    AirdropTaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        link:
          type: string
        taskType:
          type: string
          enum: [follow_twitter, join_telegram, join_discord, complete_quiz, stake, bridge, swap, provide_liquidity, register, other]
        isRequired:
          type: boolean
        sortOrder:
          type: integer
...